// src/history/types/action.type.ts
export enum ActionType {
  PRODUCT_CREATED = 'product_created',
  STOCK_CREATED = 'stock_created',
  STOCK_ON_SHELF_INCREASED = 'stock_on_shelf_increased', //увелечиние
  STOCK_ON_SHELF_DECREASED = 'stock_on_shelf_decreased', //уменьшение
  STOCK_IN_ORDER_INCREASED = 'stock_in_order_increased',
  STOCK_IN_ORDER_DECREASED = 'stock_in_order_decreased',
}